Cloud-Run/sec-filling-backfill/requirements.txt
google-cloud-bigquery>=3.0.0
requests>=2.25.0
flask>=2.2.0
functions-framework>=3.0.0
